Move AccountStatusLevelType.
